improved structure on repository

+++ RepositoryPattern/Program.cs
@@ -4,22 +4,11 @@
 // Advantage: minimizes duplicate query logic
 // Decouples your application from persistence frameworks -> Entity Framework
 using RepositoryPattern;
-using RepositoryPattern.Model;
+using System;
+using System.Linq;
 
 // The original idea envolved using DTOs but got removed to avoid complexity
 var personRepository = new PersonRepository();
-
-// Create List of People for Mockup Data
-var people = new List<Person>(){
-    new Person { Name = "Joe", Age = 10 },
-    new Person { Name = "Mike", Age = 15 },
-    new Person { Name = "John", Age = 18 },
-    new Person { Name = "Monique", Age = 12 },
-    new Person { Name = "Pedro", Age = 20 }
-};
-
-//Add Them to repository
-personRepository.AddRange(people);
 
 Console.WriteLine("List Of People\n");
 personRepository.GetAll().ToList().ForEach(Console.WriteLine);
@@ -28,7 +17,10 @@ var youngestPerson =  personRepository.GetYoungestPerson();
 Console.WriteLine("\nYoungest:\n");
 Console.WriteLine(youngestPerson);
 
-personRepository.Remove(youngestPerson);
+if(youngestPerson != null)
+{
+    personRepository.Remove(youngestPerson);
+}
 
 var Top5Youngest = personRepository.GetTopYoungest(5).ToList();
 Console.WriteLine("\nTop 5 Youngest:\n");
